Set wildtype DASM predictions to nan (#100)
Adds a new model method `predictions_of_sequences` which in the DASM case (models with output dimension >= 20) modifies the output of the existing `evaluate_sequences` method so that wildtype predictions are set to nan.
`Crepe.__call__` now calls this new method, and docstrings explain that the wildtype predictions are meaningless.