Paired sequences validation companion PR (#108)
Enables validation on light chain sequences, using the OEPlotter framework already in `dnsm-experiments-1`.
Most changes are in https://github.com/matsengrp/dnsm-experiments-1/pull/84
The netam component of this is simple: a function that computes a heavy chain and light chain mask from an aa index tensor representing a heavy/light sequence pair. This PR includes that function and a test.